refactor ResultsList and resolve VQA issues
