+++ src/components/EditionDetail/Edition.tsx
@@ -20,15 +20,12 @@ import EditionDetailDefinitionList from "~/src/components/EditionDetailDefinitio
 import Link from "~/src/components/Link/Link";
 import { InstanceCard } from "../InstanceCard/InstanceCard";
 import SearchHeader from "../SearchHeader/SearchHeader";
-import {
-  joinArrayOfElements,
-  truncateStringOnWhitespace,
-} from "~/src/util/Util";
+import { truncateStringOnWhitespace } from "~/src/util/Util";
 import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
 import { Instance } from "~/src/types/DataModel";
-import EditionCardUtils from "~/src/util/EditionCardUtils";
 import { PLACEHOLDER_LINK } from "~/src/constants/collection";
 import DrbBreakout from "../DrbBreakout/DrbBreakout";
+import AuthorsList from "../AuthorsList/AuthorsList";
 
 const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
   props
@@ -39,7 +36,6 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
   const { pathname, query } = router;
   const featuredItemId = query.featured as string;
   const edition: ApiEdition = props.editionResult.data;
-  const authorsList = EditionCardUtils.getAuthorsList(edition.work_authors);
 
   const passedInFeaturedItem = featuredItemId
     ? edition.instances.find((instance) => {
@@ -119,8 +115,10 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
         )}
       </Flex>
       {edition.sub_title && <Box>{edition.sub_title}</Box>}
-      {authorsList && authorsList.length && (
-        <Box>By {joinArrayOfElements(authorsList, "")}</Box>
+      {edition.work_authors && edition.work_authors.length && (
+        <Box>
+          By <AuthorsList authors={edition.work_authors} />
+        </Box>
       )}
       <Box>
         {featuredInstance && (
